test(master): keep journal integration tests out of core modules
