+++ curvine-master/src/master/journal/journal_system.rs
@@ -360,6 +360,13 @@ impl JournalSystem {
         self.raft_journal.app_store().clone()
     }
 
+    #[doc(hidden)]
+    pub fn append_committed_entry_for_test(&self, entry: Entry) -> RaftResult<()> {
+        let index = entry.index;
+        self.raft_journal.log_store().append(&[entry])?;
+        self.raft_journal.log_store().set_hard_state_commit(index)
+    }
+
     pub fn state_listener(&self) -> RoleStateListener {
         self.raft_journal.new_state_listener()
     }
@@ -486,108 +493,6 @@ mod tests {
             );
         }
 
-        Ok(())
-    }
-
-    #[test]
-    fn promotion_applies_committed_metadata_before_returning() -> FsResult<()> {
-        Master::init_test_metrics();
-
-        let mut source_conf = ClusterConf {
-            testing: true,
-            ..Default::default()
-        };
-        source_conf.change_test_meta_dir(format!("promotion-source-{}", Utils::rand_str(6)));
-        let source_fs = JournalSystem::fs_only_for_test(&source_conf)?;
-        source_fs.mkdir("/committed-before-promotion", false)?;
-        let entry = source_fs
-            .fs_dir
-            .read()
-            .take_entries()
-            .into_iter()
-            .next()
-            .expect("mkdir must emit a journal entry");
-
-        let mut target_conf = ClusterConf {
-            testing: true,
-            ..Default::default()
-        };
-        target_conf.change_test_meta_dir(format!("promotion-target-{}", Utils::rand_str(6)));
-        let target = JournalSystem::from_conf(&target_conf)?;
-        let target_fs = target.fs();
-        let loader = target.journal_loader();
-
-        let mut batch = JournalBatch::new(1);
-        batch.push(entry);
-        let raft_entry = Entry {
-            term: 1,
-            index: 1,
-            data: SerdeUtils::serialize(&batch)?,
-            ..Default::default()
-        };
-        target.raft_journal.log_store().append(&[raft_entry])?;
-        target.raft_journal.log_store().set_hard_state_commit(1)?;
-
-        AsyncRuntime::single().block_on(async { loader.role_change(StateRole::Leader).await })?;
-
-        assert!(target_fs.file_status("/committed-before-promotion").is_ok());
-        Ok(())
-    }
-
-    #[test]
-    fn replay_rejects_duplicate_allocated_inode_id() -> FsResult<()> {
-        Master::init_test_metrics();
-
-        let mut source_conf = ClusterConf {
-            testing: true,
-            ..Default::default()
-        };
-        source_conf.change_test_meta_dir(format!("duplicate-id-source-{}", Utils::rand_str(6)));
-        let source_fs = JournalSystem::fs_only_for_test(&source_conf)?;
-        source_fs.mkdir("/first", false)?;
-        source_fs.mkdir("/second", false)?;
-        let mut entries = source_fs.fs_dir.read().take_entries();
-        let first = entries.remove(0);
-        let mut second = entries.remove(0);
-        let duplicated_id = match &first {
-            JournalEntry::Mkdir(entry) => entry.dir.id,
-            _ => unreachable!("mkdir must emit a Mkdir journal entry"),
-        };
-        match &mut second {
-            JournalEntry::Mkdir(entry) => entry.dir.id = duplicated_id,
-            _ => unreachable!("mkdir must emit a Mkdir journal entry"),
-        }
-
-        let mut target_conf = ClusterConf {
-            testing: true,
-            ..Default::default()
-        };
-        target_conf.change_test_meta_dir(format!("duplicate-id-target-{}", Utils::rand_str(6)));
-        let target = JournalSystem::from_conf(&target_conf)?;
-        let loader = target.journal_loader();
-
-        let mut first_batch = JournalBatch::new(1);
-        first_batch.push(first);
-        let first_entry = Entry {
-            term: 1,
-            index: 1,
-            data: SerdeUtils::serialize(&first_batch)?,
-            ..Default::default()
-        };
-        let mut second_batch = JournalBatch::new(2);
-        second_batch.push(second);
-        let second_entry = Entry {
-            term: 1,
-            index: 2,
-            data: SerdeUtils::serialize(&second_batch)?,
-            ..Default::default()
-        };
-
-        let result = AsyncRuntime::single().block_on(async {
-            loader.apply(true, ApplyMsg::new_entry(first_entry)).await?;
-            loader.apply(true, ApplyMsg::new_entry(second_entry)).await
-        });
-        assert!(result.is_err());
         Ok(())
     }
 }
